files for keysets moved within packages
