+++ android/src/main/java/io/parity/signer/ui/navigationselectors/KeySetDetailsNavSubgraph.kt
@@ -9,8 +9,8 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import io.parity.signer.models.*
-import io.parity.signer.screens.keysets.details.KeySetDetailsMultiselectScreen
-import io.parity.signer.screens.keysets.details.KeySetDetailsScreenFull
+import io.parity.signer.screens.keysetdetails.export.KeySetDetailsMultiselectScreen
+import io.parity.signer.screens.keysetdetails.KeySetDetailsScreenFull
 
 @Composable
 fun KeySetDetailsNavSubgraph(
