+++ android/src/main/java/io/parity/signer/ui/navigationselectors/SignerSelectorsLegacy.kt
@@ -9,8 +9,8 @@ import io.parity.signer.bottomsheets.*
 import io.parity.signer.components.Documents
 import io.parity.signer.models.*
 import io.parity.signer.screens.*
-import io.parity.signer.screens.keysets.*
-import io.parity.signer.screens.keysets.details.SeedMenu
+import io.parity.signer.screens.keysetdetails.old.KeyManager
+import io.parity.signer.screens.keysetdetails.old.SeedMenu
 import io.parity.signer.uniffi.Action
 import io.parity.signer.uniffi.AlertData
 import io.parity.signer.uniffi.ModalData
